Update the previous CustomAccordion class to be a full web component utilizing a shadow dom.
The original class was overly complex and awkward to parse and read. This new one is fairly simple, although it requires whatever code using it to dynamically altar the button trigger text and accordion content using slot elements.

The reason for this change was my own personal interest in exploring how web components work. They seem like a powerful way to encapsulate shared HTML, but styling seems awkward, and I have ignore it for now.