Add caching for RSS Feed items
Caching strategy is a naive implementation - it simply relies on a expiration time for evicting feeds from the cache.
RSS Feed items are inserted via stringify into the cache. Insertion occurs when the RSS Feed is first added to the DB and when the specific id is requested.

The expiration time is a bit tricky to figure out. RSS spec mentions an optional ttl element that may be used. The cache control header is also available for cache timing.
So the strategy here is to set a default cache time of one hour. If the ttl element exists, then we convert that from minutes to seconds and use it. If a cache control header is present, that take precedence unless its 0.

We also delete the feed from the cache if the feed is deleted.